requirements.txt
alabaster==0.7.10
alembic==0.8.9
astroid==1.6.3
attrs==17.4.0
Babel==2.5.3
bcrypt==3.1.2
beautifulsoup4==4.6.0
bleach==1.5.0
boto==2.48.0
boto3==1.5.8
botocore==1.8.22
bs4==0.0.1
bz2file==0.98
certifi==2017.11.5
cffi==1.9.1
chardet==3.0.4
click==6.7
coloredlogs==8.0
colorlog==3.1.0
coverage==4.2
docutils==0.14
enum34==1.1.6
Flask==0.12.2
Flask-API==1.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-HTTPAuth==3.2.3
Flask-JWT==0.3.2
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
gensim==3.2.0
gunicorn==19.7.1
h5py==2.7.1
html5lib==0.9999999
httpie==0.9.9
humanfriendly==4.6
idna==2.6
imagesize==1.0.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
Keras==2.1.5
lazy-object-proxy==1.3.1
Mako==1.0.6
Markdown==2.6.9
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.14.2
numpydoc==0.8.0
olefile==0.44
opencv-python==3.3.0.10
packaging==17.1
pandas==0.22.0
Pillow==4.3.0
pluggy==0.6.0
protobuf==3.4.0
psycopg2==2.7.3.2
py==1.5.2
pycparser==2.17
Pygments==2.2.0
PyJWT==1.4.2
pylint==1.8.4
pyparsing==2.2.0
pytest==3.3.2
pytest-flask==0.10.0
python-dateutil==2.6.1
python-editor==1.0.3
python-mimeparse==1.6.0
pytz==2017.3
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
smart-open==1.5.6
snowballstemmer==1.2.1
Sphinx==1.7.2
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.1.4
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
typing==3.6.4
urllib3==1.22
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
